refactor: Move command architecture files
