+++ src/commands/slashcommands/create-difficulty.ts
@@ -2,13 +2,13 @@ import { SlashCommandBuilder } from '@discordjs/builders';
 import { CommandInteractionOption, GuildMember, PermissionsBitField } from 'discord.js';
 import { createDifficultyInTournament, getDifficultyByEmoji, getTournamentByName, isSingleEmoji } from '../../backend/queries/tournamentQueries.js';
 import { getCurrentTournament } from '../../backend/queries/guildSettingsQueries.js';
-import { SimpleRendezvousSlashCommand } from './architecture/rendezvousCommand.js';
+import { SimpleRendezvousSlashCommand } from '../architecture/rendezvousCommand.js';
 import { OptionValidationErrorOutcome, Outcome, OutcomeStatus, SlashCommandDescribedOutcome } from '../../types/outcome.js';
 import { LimitedCommandInteraction } from '../../types/limitedCommandInteraction.js';
 import { getJudgeByGuildIdAndMemberId } from '../../backend/queries/profileQueries.js';
 import { OptionValidationError, OptionValidationErrorStatus } from '../../types/customError.js';
 import { ValueOf } from '../../types/typelogic.js';
-import { Constraint, validateConstraints } from './architecture/validation.js';
+import { Constraint, validateConstraints } from '../architecture/validation.js';
 
 /**
  * Alias for the first generic type of the command.
